+++ models/wide_resnet_cifar.py
@@ -8,7 +8,7 @@ import torch
 import torch.nn as nn
 from torch.autograd import Variable
 import math
-from resnet_cifar import BasicBlock
+from .resnet_cifar import BasicBlock
 
 
 class Wide_ResNet_Cifar(nn.Module):
@@ -67,7 +67,7 @@ class Wide_ResNet_Cifar(nn.Module):
 
 def wide_resnet_cifar(depth, width, **kwargs):
     assert (depth - 2) % 6 == 0
-    n = (depth - 2) / 6
+    n = int( (depth - 2) / 6 )
     return Wide_ResNet_Cifar(BasicBlock, [n, n, n], width, **kwargs)
